+++ packages/e2e-tests/prepare.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+export $(cat ./.env | grep -v '#' | xargs)
+pushd ./hardhat
+npm run build:docker
+popd
+
+pushd ../subtree-updater
+yarn build:mock:docker
+popd 
+
+pushd ../bundler
+yarn build:docker
+popd